+++ docs/changelog.rst
@@ -27,7 +27,9 @@ Chula v0.9.0 (dev)
   [crude] stack trace formatting in html when debugging is enabled.
 - Added support for :attr:`chula.config.Config.auto_reload` to make coding
   even funner.
-- Improved logging.
+- [`GH-16 <http://github.com/jmcfarlane/chula/issues#issue/16>`_]
+  Improved logging.  Both error and debug logging use files.  This
+  fixes the long standing stderr blocking issue.
 - Improved build process to reference build artifacts directly from
   pypi.
 - [`GH-20 <http://github.com/jmcfarlane/chula/issues#issue/20>`_]